Remove / Consolidate legacy usage adoption metrics (#100480)
* Remove legacydetection rule stat summaries

* Remove ML usage summary and consolidate with ML metric telemetry.

* Remove ML usage summary and consolidate with ML metric telemetry.

* Move legacy helper constructs into index.

* Separate rule logic from ml logic. Add ml unit tests.

* Abstract types away into their own file.

* Update telemetry schema.

# Conflicts:
#	x-pack/plugins/security_solution/server/usage/collector.ts
#	x-pack/plugins/security_solution/server/usage/detections/index.ts
#	x-pack/plugins/telemetry_collection_xpack/schema/xpack_plugins.json
