+++ x-pack/plugins/security_solution/server/usage/detections/detection_rule_helpers.ts
@@ -5,22 +5,18 @@
  * 2.0.
  */
 
-import {
-  ElasticsearchClient,
-  KibanaRequest,
-  SavedObjectsClientContract,
-} from '../../../../../../src/core/server';
+import { ElasticsearchClient, SavedObjectsClientContract } from '../../../../../../src/core/server';
+import { SIGNALS_ID } from '../../../common/constants';
+import { isElasticRule } from './index';
 import {
   AlertsAggregationResponse,
   CasesSavedObject,
   DetectionRulesTypeUsage,
   DetectionRuleMetric,
   DetectionRuleAdoption,
-  MlJobMetric,
-} from './index';
-import { SIGNALS_ID } from '../../../common/constants';
-import { DatafeedStats, Job, MlPluginSetup } from '../../../../ml/server';
-import { isElasticRule, RuleSearchParams, RuleSearchResult } from './detection_telemetry_helpers';
+  RuleSearchParams,
+  RuleSearchResult,
+} from './types';
 
 /**
  * Default detection rule usage count, split by type + elastic/custom
@@ -287,92 +283,4 @@ export const getDetectionRuleMetrics = async (
     detection_rule_detail: [],
     detection_rule_usage: rulesUsage,
   };
-};
-
-export const getMlJobMetrics = async (
-  ml: MlPluginSetup | undefined,
-  savedObjectClient: SavedObjectsClientContract
-): Promise<MlJobMetric[]> => {
-  if (ml) {
-    try {
-      const fakeRequest = { headers: {} } as KibanaRequest;
-      const jobsType = 'security';
-      const securityJobStats = await ml
-        .anomalyDetectorsProvider(fakeRequest, savedObjectClient)
-        .jobStats(jobsType);
-
-      const jobDetails = await ml
-        .anomalyDetectorsProvider(fakeRequest, savedObjectClient)
-        .jobs(jobsType);
-
-      const jobDetailsCache = new Map<string, Job>();
-      jobDetails.jobs.forEach((detail) => jobDetailsCache.set(detail.job_id, detail));
-
-      const datafeedStats = await ml
-        .anomalyDetectorsProvider(fakeRequest, savedObjectClient)
-        .datafeedStats();
-
-      const datafeedStatsCache = new Map<string, DatafeedStats>();
-      datafeedStats.datafeeds.forEach((datafeedStat) =>
-        datafeedStatsCache.set(`${datafeedStat.datafeed_id}`, datafeedStat)
-      );
-
-      return securityJobStats.jobs.map((stat) => {
-        const jobId = stat.job_id;
-        const jobDetail = jobDetailsCache.get(stat.job_id);
-        const datafeed = datafeedStatsCache.get(`datafeed-${jobId}`);
-
-        return {
-          job_id: jobId,
-          open_time: stat.open_time,
-          create_time: jobDetail?.create_time,
-          finished_time: jobDetail?.finished_time,
-          state: stat.state,
-          data_counts: {
-            bucket_count: stat.data_counts.bucket_count,
-            empty_bucket_count: stat.data_counts.empty_bucket_count,
-            input_bytes: stat.data_counts.input_bytes,
-            input_record_count: stat.data_counts.input_record_count,
-            last_data_time: stat.data_counts.last_data_time,
-            processed_record_count: stat.data_counts.processed_record_count,
-          },
-          model_size_stats: {
-            bucket_allocation_failures_count:
-              stat.model_size_stats.bucket_allocation_failures_count,
-            memory_status: stat.model_size_stats.memory_status,
-            model_bytes: stat.model_size_stats.model_bytes,
-            model_bytes_exceeded: stat.model_size_stats.model_bytes_exceeded,
-            model_bytes_memory_limit: stat.model_size_stats.model_bytes_memory_limit,
-            peak_model_bytes: stat.model_size_stats.peak_model_bytes,
-          },
-          timing_stats: {
-            average_bucket_processing_time_ms: stat.timing_stats.average_bucket_processing_time_ms,
-            bucket_count: stat.timing_stats.bucket_count,
-            exponential_average_bucket_processing_time_ms:
-              stat.timing_stats.exponential_average_bucket_processing_time_ms,
-            exponential_average_bucket_processing_time_per_hour_ms:
-              stat.timing_stats.exponential_average_bucket_processing_time_per_hour_ms,
-            maximum_bucket_processing_time_ms: stat.timing_stats.maximum_bucket_processing_time_ms,
-            minimum_bucket_processing_time_ms: stat.timing_stats.minimum_bucket_processing_time_ms,
-            total_bucket_processing_time_ms: stat.timing_stats.total_bucket_processing_time_ms,
-          },
-          datafeed: {
-            datafeed_id: datafeed?.datafeed_id,
-            state: datafeed?.state,
-            timing_stats: {
-              bucket_count: datafeed?.timing_stats.bucket_count,
-              exponential_average_search_time_per_hour_ms:
-                datafeed?.timing_stats.exponential_average_search_time_per_hour_ms,
-              search_count: datafeed?.timing_stats.search_count,
-              total_search_time_ms: datafeed?.timing_stats.total_search_time_ms,
-            },
-          },
-        } as MlJobMetric;
-      });
-    } catch (e) {
-      // ignore failure, usage will be zeroed
-    }
-  }
-
-  return [];
 };
