+++ x-pack/plugins/security_solution/server/usage/detections/detections.test.ts
@@ -11,10 +11,11 @@ import {
   savedObjectsClientMock,
 } from '../../../../../../src/core/server/mocks';
 import { mlServicesMock } from '../../lib/machine_learning/mocks';
+import { fetchDetectionsMetrics } from './index';
+import { initialMlJobsUsage } from './detection_ml_helpers';
 import {
   getMockJobSummaryResponse,
   getMockListModulesResponse,
-  getMockRulesResponse,
   getMockMlJobDetailsResponse,
   getMockMlJobStatsResponse,
   getMockMlDatafeedStatsResponse,
@@ -22,96 +23,12 @@ import {
   getMockRuleAlertsResponse,
   getMockAlertCasesResponse,
 } from './detections.mocks';
-import { fetchDetectionsUsage, fetchDetectionsMetrics } from './index';
 
 const savedObjectsClient = savedObjectsClientMock.create();
 
 describe('Detections Usage and Metrics', () => {
   let esClientMock: jest.Mocked<ElasticsearchClient>;
   let mlMock: ReturnType<typeof mlServicesMock.createSetupContract>;
-
-  describe('fetchDetectionsUsage()', () => {
-    beforeEach(() => {
-      esClientMock = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      mlMock = mlServicesMock.createSetupContract();
-    });
-
-    it('returns zeroed counts if both calls are empty', async () => {
-      const result = await fetchDetectionsUsage('', esClientMock, mlMock, savedObjectsClient);
-
-      expect(result).toEqual({
-        detection_rules: {
-          custom: {
-            enabled: 0,
-            disabled: 0,
-          },
-          elastic: {
-            enabled: 0,
-            disabled: 0,
-          },
-        },
-        ml_jobs: {
-          custom: {
-            enabled: 0,
-            disabled: 0,
-          },
-          elastic: {
-            enabled: 0,
-            disabled: 0,
-          },
-        },
-      });
-    });
-
-    it('tallies rules data given rules results', async () => {
-      (esClientMock.search as jest.Mock).mockResolvedValue({ body: getMockRulesResponse() });
-
-      const result = await fetchDetectionsUsage('', esClientMock, mlMock, savedObjectsClient);
-
-      expect(result).toEqual(
-        expect.objectContaining({
-          detection_rules: {
-            custom: {
-              enabled: 1,
-              disabled: 1,
-            },
-            elastic: {
-              enabled: 2,
-              disabled: 3,
-            },
-          },
-        })
-      );
-    });
-
-    it('tallies jobs data given jobs results', async () => {
-      const mockJobSummary = jest.fn().mockResolvedValue(getMockJobSummaryResponse());
-      const mockListModules = jest.fn().mockResolvedValue(getMockListModulesResponse());
-      mlMock.modulesProvider.mockReturnValue(({
-        listModules: mockListModules,
-      } as unknown) as ReturnType<typeof mlMock.modulesProvider>);
-      mlMock.jobServiceProvider.mockReturnValue({
-        jobsSummary: mockJobSummary,
-      });
-
-      const result = await fetchDetectionsUsage('', esClientMock, mlMock, savedObjectsClient);
-
-      expect(result).toEqual(
-        expect.objectContaining({
-          ml_jobs: {
-            custom: {
-              enabled: 1,
-              disabled: 1,
-            },
-            elastic: {
-              enabled: 1,
-              disabled: 1,
-            },
-          },
-        })
-      );
-    });
-  });
 
   describe('getDetectionRuleMetrics()', () => {
     beforeEach(() => {
@@ -171,7 +88,7 @@ describe('Detections Usage and Metrics', () => {
               },
             },
           },
-          ml_jobs: [],
+          ml_jobs: { ml_job_metrics: [], ml_job_usage: initialMlJobsUsage },
         })
       );
     });
@@ -246,7 +163,7 @@ describe('Detections Usage and Metrics', () => {
               },
             },
           },
-          ml_jobs: [],
+          ml_jobs: { ml_job_metrics: [], ml_job_usage: initialMlJobsUsage },
         })
       );
     });
@@ -308,7 +225,7 @@ describe('Detections Usage and Metrics', () => {
               },
             },
           },
-          ml_jobs: [],
+          ml_jobs: { ml_job_metrics: [], ml_job_usage: initialMlJobsUsage },
         })
       );
     });
@@ -383,7 +300,7 @@ describe('Detections Usage and Metrics', () => {
               },
             },
           },
-          ml_jobs: [],
+          ml_jobs: { ml_job_metrics: [], ml_job_usage: initialMlJobsUsage },
         })
       );
     });
@@ -404,12 +321,20 @@ describe('Detections Usage and Metrics', () => {
 
       expect(result).toEqual(
         expect.objectContaining({
-          ml_jobs: [],
+          ml_jobs: { ml_job_metrics: [], ml_job_usage: initialMlJobsUsage },
         })
       );
     });
 
     it('returns an ml job telemetry object from anomaly detectors provider', async () => {
+      const mockJobSummary = jest.fn().mockResolvedValue(getMockJobSummaryResponse());
+      const mockListModules = jest.fn().mockResolvedValue(getMockListModulesResponse());
+      mlMock.modulesProvider.mockReturnValue(({
+        listModules: mockListModules,
+      } as unknown) as ReturnType<typeof mlMock.modulesProvider>);
+      mlMock.jobServiceProvider.mockReturnValue({
+        jobsSummary: mockJobSummary,
+      });
       const mockJobsResponse = jest.fn().mockResolvedValue(getMockMlJobDetailsResponse());
       const mockJobStatsResponse = jest.fn().mockResolvedValue(getMockMlJobStatsResponse());
       const mockDatafeedStatsResponse = jest
@@ -426,49 +351,61 @@ describe('Detections Usage and Metrics', () => {
 
       expect(result).toEqual(
         expect.objectContaining({
-          ml_jobs: [
-            {
-              job_id: 'high_distinct_count_error_message',
-              create_time: 1603838214983,
-              finished_time: 1611739871669,
-              state: 'closed',
-              data_counts: {
-                bucket_count: 8612,
-                empty_bucket_count: 8590,
-                input_bytes: 45957,
-                input_record_count: 162,
-                last_data_time: 1610470367123,
-                processed_record_count: 162,
+          ml_jobs: {
+            ml_job_usage: {
+              custom: {
+                disabled: 1,
+                enabled: 1,
               },
-              model_size_stats: {
-                bucket_allocation_failures_count: 0,
-                memory_status: 'ok',
-                model_bytes: 72574,
-                model_bytes_exceeded: 0,
-                model_bytes_memory_limit: 16777216,
-                peak_model_bytes: 78682,
-              },
-              timing_stats: {
-                average_bucket_processing_time_ms: 0.4900837644740133,
-                bucket_count: 16236,
-                exponential_average_bucket_processing_time_ms: 0.23614068552903306,
-                exponential_average_bucket_processing_time_per_hour_ms: 1.5551298175461634,
-                maximum_bucket_processing_time_ms: 392,
-                minimum_bucket_processing_time_ms: 0,
-                total_bucket_processing_time_ms: 7957.00000000008,
-              },
-              datafeed: {
-                datafeed_id: 'datafeed-high_distinct_count_error_message',
-                state: 'stopped',
-                timing_stats: {
-                  bucket_count: 8612,
-                  exponential_average_search_time_per_hour_ms: 86145.39799630083,
-                  search_count: 7202,
-                  total_search_time_ms: 3107147,
-                },
+              elastic: {
+                disabled: 1,
+                enabled: 1,
               },
             },
-          ],
+            ml_job_metrics: [
+              {
+                job_id: 'high_distinct_count_error_message',
+                create_time: 1603838214983,
+                finished_time: 1611739871669,
+                state: 'closed',
+                data_counts: {
+                  bucket_count: 8612,
+                  empty_bucket_count: 8590,
+                  input_bytes: 45957,
+                  input_record_count: 162,
+                  last_data_time: 1610470367123,
+                  processed_record_count: 162,
+                },
+                model_size_stats: {
+                  bucket_allocation_failures_count: 0,
+                  memory_status: 'ok',
+                  model_bytes: 72574,
+                  model_bytes_exceeded: 0,
+                  model_bytes_memory_limit: 16777216,
+                  peak_model_bytes: 78682,
+                },
+                timing_stats: {
+                  average_bucket_processing_time_ms: 0.4900837644740133,
+                  bucket_count: 16236,
+                  exponential_average_bucket_processing_time_ms: 0.23614068552903306,
+                  exponential_average_bucket_processing_time_per_hour_ms: 1.5551298175461634,
+                  maximum_bucket_processing_time_ms: 392,
+                  minimum_bucket_processing_time_ms: 0,
+                  total_bucket_processing_time_ms: 7957.00000000008,
+                },
+                datafeed: {
+                  datafeed_id: 'datafeed-high_distinct_count_error_message',
+                  state: 'stopped',
+                  timing_stats: {
+                    bucket_count: 8612,
+                    exponential_average_search_time_per_hour_ms: 86145.39799630083,
+                    search_count: 7202,
+                    total_search_time_ms: 3107147,
+                  },
+                },
+              },
+            ],
+          },
         })
       );
     });
